+++ webpack.config.js
@@ -7,7 +7,7 @@ import ResolveTypeScriptPlugin from "resolve-typescript-plugin";
 import CopyWebpackPlugin from "copy-webpack-plugin";
 import { merge } from "webpack-merge";
 import PugPlugin from "pug-plugin";
-import sass from "sass";
+import * as sass from "sass";
 
 import packageInfo from "./package.json" with { type: "json" };
 import manifestV2 from "./resources/manifest.v2.json" with { type: "json" };